character: migrate suffer() from player
Migrate suffer() down from player:: into Character::, along with the
minimum but still substantial set of functions and variables that also
need to be in Character:: in order to make the code compile.
