+++ src/character.h
@@ -38,6 +38,7 @@
 #include "string_id.h"
 #include "type_id.h"
 #include "units.h"
+#include "weighted_list.h"
 #include "point.h"
 #include "magic_enchantment.h"
 
@@ -345,6 +346,22 @@ class Character : public Creature, public visitable<Character>
         float get_dodge_base() const override;
         float get_hit_base() const override;
 
+        /**
+         * Adds a reason for why the player would miss a melee attack.
+         *
+         * To possibly be messaged to the player when he misses a melee attack.
+         * @param reason A message for the player that gives a reason for him missing.
+         * @param weight The weight used when choosing what reason to pick when the
+         * player misses.
+         */
+        void add_miss_reason( const std::string &reason, unsigned int weight );
+        /** Clears the list of reasons for why the player would miss a melee attack. */
+        void clear_miss_reasons();
+        /**
+         * Returns an explanation for why the player would miss a melee attack.
+         */
+        std::string get_miss_reason();
+
         /** Handles health fluctuations over time */
         virtual void update_health( int external_modifiers = 0 );
 
@@ -524,6 +541,7 @@ class Character : public Creature, public visitable<Character>
         void dismount();
         void forced_dismount();
 
+        bool is_deaf() const;
         /** Returns true if the player has two functioning arms */
         bool has_two_arms() const;
         /** Returns the number of functioning arms */
@@ -761,6 +779,14 @@ class Character : public Creature, public visitable<Character>
         // route for overmap-scale travelling
         std::vector<tripoint> omt_path;
 
+        /** Handles bionic effects over time of the entered bionic */
+        void process_bionic( int b );
+        /** Handles bionic deactivation effects of the entered bionic, returns if anything
+         *  deactivated */
+        virtual bool deactivate_bionic( int b, bool eff_only = false );
+        /**Convert fuel to bionic power*/
+        bool burn_fuel( int b, bool start = false );
+
         units::energy get_power_level() const;
         units::energy get_max_power_level() const;
         void mod_power_level( units::energy npower );
@@ -965,6 +991,9 @@ class Character : public Creature, public visitable<Character>
         /// used for operations on distant objects (e.g. vehicle installation/uninstallation)
         int best_nearby_lifting_assist( const tripoint &world_pos ) const;
 
+        // Inventory + weapon + worn (for death, etc)
+        std::vector<item *> inv_dump();
+
         units::mass weight_carried() const;
         units::volume volume_carried() const;
 
@@ -1012,12 +1041,26 @@ class Character : public Creature, public visitable<Character>
          * @return whether both removal and replacement were successful (they are performed atomically)
          */
         virtual bool wield( item &target ) = 0;
+        /**
+         * Check player capable of unwielding an item.
+         * @param it Thing to be unwielded
+         */
+        ret_val<bool> can_unwield( const item &it ) const;
 
         void drop_invalid_inventory();
+        /** Drops an item to the specified location */
+        void drop( int pos, const tripoint &where );
+        virtual void drop( const std::list<std::pair<int, int>> &what, const tripoint &target,
+                           bool stash = false );
 
         virtual bool has_artifact_with( art_effect_passive effect ) const;
 
         bool is_wielding( const item &target ) const;
+
+        bool covered_with_flag( const std::string &flag, const body_part_set &parts ) const;
+        bool is_waterproof( const body_part_set &parts ) const;
+        // Carried items may leak radiation or chemicals
+        int leak_level( const std::string &flag ) const;
 
         // --------------- Clothing Stuff ---------------
         /** Returns true if the player is wearing the item. */
@@ -1099,6 +1142,16 @@ class Character : public Creature, public visitable<Character>
         virtual bool query_yn( const std::string &msg ) const = 0;
 
         bool is_immune_field( field_type_id fid ) const override;
+        /** Returns true is the player is protected from electric shocks */
+        bool is_elec_immune() const override;
+        /** Returns true if the player is immune to this kind of effect */
+        bool is_immune_effect( const efftype_id & ) const override;
+        /** Returns true if the player is immune to this kind of damage */
+        bool is_immune_damage( damage_type ) const override;
+        /** Returns true if the player is protected from radiation */
+        bool is_rad_immune() const;
+        /** Returns true if the player is immune to throws */
+        bool is_throw_immune() const;
 
         /** Returns true if the player has some form of night vision */
         bool has_nv();
@@ -1182,6 +1235,23 @@ class Character : public Creature, public visitable<Character>
 
         int oxygen;
         int radiation;
+        int tank_plut;
+        int reactor_plut;
+        int slow_rad;
+
+        int focus_pool;
+        /* crafting inventory cached time */
+        time_point cached_time;
+
+        std::vector <addiction> addictions;
+        /** Adds an addiction to the player */
+        void add_addiction( add_type type, int strength );
+        /** Removes an addition from the player */
+        void rem_addiction( add_type type );
+        /** Returns true if the player has an addiction of the specified type */
+        bool has_addiction( add_type type ) const;
+        /** Returns the intensity of the specified addiction */
+        int  addiction_level( add_type type ) const;
 
         std::shared_ptr<monster> mounted_creature;
         // for loading NPC mounts
@@ -1379,7 +1449,18 @@ class Character : public Creature, public visitable<Character>
 
         /** Returns true if the player is wearing something on the entered body_part */
         bool wearing_something_on( body_part bp ) const;
-        /** Returns 1 if the player is wearing something on both feet, .5 if on one, and 0 if on neither */
+        /** Returns true if the player is wearing something occupying the helmet slot */
+        bool is_wearing_helmet() const;
+        /** Returns the total encumbrance of all SKINTIGHT and HELMET_COMPAT items coveringi
+         *  the head */
+        int head_cloth_encumbrance() const;
+        /** Same as footwear factor, but for arms */
+        double armwear_factor() const;
+        /** Returns 1 if the player is wearing an item of that count on one foot, 2 if on both,
+         *  and zero if on neither */
+        int shoe_type_count( const itype_id &it ) const;
+        /** Returns 1 if the player is wearing something on both feet, .5 if on one,
+         *  and 0 if on neither */
         double footwear_factor() const;
         /** Returns true if the player is wearing something on their feet that is not SKINTIGHT */
         bool is_wearing_shoes( const side &which_side = side::BOTH ) const;
@@ -1400,11 +1481,31 @@ class Character : public Creature, public visitable<Character>
         void clear_morale();
         bool has_morale_to_read() const;
         bool has_morale_to_craft() const;
+
+        const inventory &crafting_inventory( const tripoint &src_pos = tripoint_zero,
+                                             int radius = PICKUP_RANGE );
+        void invalidate_crafting_inventory();
+
         /** Checks permanent morale for consistency and recovers it when an inconsistency is found. */
         void check_and_recover_morale();
 
         /** Handles the enjoyability value for a comestible. First value is enjoyability, second is cap. **/
         std::pair<int, int> fun_for( const item &comest ) const;
+
+        /** Handles a large number of timers decrementing and other randomized effects */
+        void suffer();
+        /** Handles mitigation and application of radiation */
+        bool irradiate( float rads, bool bypass = false );
+        /** Handles the chance for broken limbs to spontaneously heal to 1 HP */
+        void mend( int rate_multiplier );
+
+        /** Creates an auditory hallucination */
+        void sound_hallu();
+
+        /** Drenches the player with water, saturation is the percent gotten wet */
+        void drench( int saturation, const body_part_set &flags, bool ignore_waterproof );
+        /** Recalculates morale penalty/bonus from wetness based on mutations, equipment and temperature */
+        void apply_wetness_morale( int temperature );
 
     protected:
         Character();
@@ -1521,6 +1622,8 @@ class Character : public Creature, public visitable<Character>
 
         int stim;
         int pkill;
+
+        struct weighted_int_list<std::string> melee_miss_reasons;
 
     protected:
         /** Amount of time the player has spent in each overmap tile. */
